test: delegator rewards folder structure
