+++ e2e-tests/pytest.ini
@@ -5,7 +5,7 @@ markers =
     candidate_status: active or inactive, used in test_registrations.py
     ariadne: run tests for Ariadne feature
     probability: run non deterministic tests
-    block_reward: run tests for Block Rewards feature
+    delegator_rewards: run tests for ADA Delegator Rewards feature
     registration: run tests for Registration feature (permissioned and trustless)
     committee_distribution: run tests for Committee Distribution feature
     committee_rotation: run tests for Committee Rotation feature
