+++ e2e-tests/src/blockchain_api.py
@@ -399,7 +399,9 @@ class BlockchainApi(ABC):
         pass
 
     @abstractmethod
-    def sign_block_producer_metadata(self, metadata: dict, cross_chain_signing_key: str) -> BlockProducerMetadataSignature:
+    def sign_block_producer_metadata(
+        self, metadata: dict, cross_chain_signing_key: str
+    ) -> BlockProducerMetadataSignature:
         """
         Creates a signature for block procuder metadata.
 
@@ -485,7 +487,7 @@ class BlockchainApi(ABC):
         """
         Calls testHelperPallet for block participation data. This helper pallet returns raw inherent data that can be
         used by chain builders to implement rewards distribution logic.
-        Helper pallet releases data every 30 slots.
+        Helper pallet releases data in a block produced in a slot divisible by 30.
 
         Arguments:
             block_hash {str} -- PC block hash
