+++ e2e-tests/tests/delegator_rewards/test_spo_journey.py
@@ -9,13 +9,22 @@ from pytest import fixture, mark, skip
 PARTICIPATION_DATA_SLOT_RANGE = 30
 
 
+##############################################################################################################
+# Disclaimer!                                                                                                #
+# The Registered SPO user journey is not fully covered because it requires a rewards distribution mechanism. #
+# Its implementation lies on the Chain Builder side. This test suite is designed to verify the correctness   #
+# of the raw inherent data that we pass to the Chain Builder to build such a mechanism. The raw inherent     #
+# data are exposed via testHelperPallet which serves as a replacement for distribution mechanism.            #
+##############################################################################################################
+
+
 @fixture(scope="module")
 def block_to_query_storage(api: BlockchainApi):
     block = api.get_block()
     block_no = block["header"]["number"]
     if block_no <= PARTICIPATION_DATA_SLOT_RANGE:
         skip(f"Participation data is released after {PARTICIPATION_DATA_SLOT_RANGE} slots, current block {block_no}.")
-    logging.info(f"Block to query storage: {block['header']['number']}")
+    logging.info(f"Block to query storage: {block_no}")
     return block
 
 
@@ -43,10 +52,10 @@ def pc_epochs(block_participation, config: ApiConfig, initial_pc_epoch):
     logging.info(f"Participation data up to slot: {up_to_slot}")
     start_pc_epoch = (up_to_slot - PARTICIPATION_DATA_SLOT_RANGE) // config.nodes_config.slots_in_epoch
     stop_pc_epoch = up_to_slot // config.nodes_config.slots_in_epoch
-    epochs = range(start_pc_epoch, stop_pc_epoch + 1)
+    epochs = range(start_pc_epoch, stop_pc_epoch)
     logging.info(f"Participation data spans PC epochs: {epochs}")
     if initial_pc_epoch in epochs:
-        epochs = range(initial_pc_epoch, stop_pc_epoch + 1)
+        epochs = range(initial_pc_epoch, stop_pc_epoch)
         logging.info(f"Initial PC epoch is greater than the first PC epoch. Adjusting... New range is: {epochs}")
     return epochs
 
@@ -81,9 +90,9 @@ def count_blocks(pc_epoch_calculator: PartnerChainEpochCalculator, config: ApiCo
             return mc_epoch_to_pc_slots[mc_epoch]
 
         pc_epochs = pc_epoch_calculator.find_pc_epochs(mc_epoch, start_from_initial_pc_epoch=False)
-        first_slot = pc_epochs.start * slots_in_epoch
-        last_slot = pc_epochs.stop * slots_in_epoch
-        slots = range(first_slot, last_slot)
+        start_slot = pc_epochs.start * slots_in_epoch
+        stop_slot = pc_epochs.stop * slots_in_epoch
+        slots = range(start_slot, stop_slot)
         logging.info(f"MC epoch {mc_epoch} PC slots: {slots}")
         mc_epoch_to_pc_slots[mc_epoch] = slots
         return slots
@@ -133,7 +142,6 @@ def test_pro_bono_participation(
             if expected_producer["block_count"] == 0:
                 logging.info(f"No blocks produced by ProBono producer {permissioned_candidate['sidechainPublicKey']}")
                 continue
-            # expected_producer["mc_epoch"] = mc_epoch - 2
             expected_producer["delegator_total_shares"] = 0
             expected_producer["delegators"] = []
             logging.info(f"Expected ProBono Producer: {expected_producer}")
@@ -163,7 +171,6 @@ def test_spo_participation(
                 continue
 
             mc_epoch_for_stake = mc_epoch - 2
-            # expected_spo["mc_epoch"] = mc_epoch_for_stake
             stake_pool_id = api.cardano_cli.get_stake_pool_id(cold_vkey=mc_pub_key[2:], output_format="bech32")
             query = text(
                 "SELECT sa.view AS stake_address, encode(sa.hash_raw, 'hex') AS stake_hash, es.amount AS stake_amount "
